# Optimize generating mock DataFrames