## What do these changes do?

Optimize generating mock DataFrames when num of columns is large.

Test results:

```python
import mars.tensor as mt, mars.dataframe as md
a = mt.random.randint(low=1, high=1000, size=(20000, 40000))
```

```python
%%time
df = md.DataFrame(a)
_ = df.min()
```

```
Before:
CPU times: user 28.2 s, sys: 302 ms, total: 28.5 s
Wall time: 28.5 s
After: 
CPU times: user 4.44 s, sys: 55.4 ms, total: 4.5 s
Wall time: 4.52 s
```

```python
from mars.dataframe.utils import build_empty_df
%time _ = build_empty_df(df.dtypes)
```

```
Before:
CPU times: user 28.4 s, sys: 297 ms, total: 28.7 s
Wall time: 28.7 s
After:
CPU times: user 4.36 s, sys: 39 ms, total: 4.4 s
Wall time: 4.4 s
```

## Related issue number

<!-- Are there any issues opened that will be resolved by merging this change? -->